feat: Refactor and improve classes, including language update

+++ src/main/java/n1exercici1/Product.java
@@ -3,14 +3,14 @@ package n1exercici1;
 public class Product {
 
     private String name;
-    private int price;
+    private double price;
 
-    public Product(String name, int price) {
+    public Product(String name, double price) {
         this.name = name;
         this.price = price;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return this.price;
     }
